ona-edge/requirements.txt
# ONA Edge Requirements

# DICOM
pydicom>=2.4.0
pynetdicom>=2.0.0

# AI/ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Web/API
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
websockets>=11.0.0

# Database
aiosqlite>=0.19.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
